Created Piece.java
I made the player variable a char since I think we can use 'w' and 'b' for now, but if it needs to be more complex we can definitely make a Player class. I also separated checking if it is a valid move and making the move for now, but we can combine it if its unnecessary.